Changed AidHui.flags type from float to uint32_t so bitwise logic in net_assist.cpp will work

+++ include/ublox/ublox_structures.h
@@ -257,7 +257,7 @@ PACK(
         float klobb1;
         float klobb2;
         float klobb3;
-        float flags;
+        uint32_t flags;
         uint8_t checksum[2];
 });
